+++ NEWS.md
@@ -1,10 +1,18 @@
-# NEWS
+# Sovereign: State-Dependent Empirical Analysis  
 
 ## Version 1.0.1 
 
-- Include COVID shock correction a la Lenza and Primiceri (2020)
-- Include historical decomposition of VAR shocks  
-- Specify structural assumptions in IRFs  
+- Major updates
+    - Include COVID shock correction a la Lenza and Primiceri (2020)
+    - Include historical decomposition of VAR shocks
+      
+- Minor updates
+    - Create object classes for LP and VAR output  
+    - Specify structural assumptions in IRFs  
+    - n.lag no longer requires a 'date' column 
+
+- Bug fixes
+    - Convert errorConditiont() to stop()   
 
 ## Version 1.0.0 
 (2021-04-02)